requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
python-docx
Pillow
google-generativeai
requests
beautifulsoup4